give-me-pizza/src/main/java/br/com/give_me_pizza/Customer.java
package br.com.give_me_pizza;

public class Customer {
	private int customerID;
	private String firstName;
	private String surname;
	private Address address;

	public String getFirstName() {
		return firstName;
	}

	public void setFirstName(String firstName) {
		this.firstName = firstName;
	}

	public String getSurname() {
		return surname;
	}

	public void setSurname(String surname) {
		this.surname = surname;
	}

	public int getCustomerID() {
		return customerID;
	}

	public void setCustomerID(int customerID) {
		this.customerID = customerID;
	}

	public Address getAddress() {
		return address;
	}

	public void setAddress(Address address) {
		this.address = address;
	}

	public String toString() {
		return (this.getFirstName() + " " + this.getSurname() + "\n" + this.getAddress().toString());
	}
}
